backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
cryptography==45.0.6
tronpy==0.5.0
redis==6.4.0
rq==2.5.0
python-multipart==0.0.20
httpx==0.28.1
openpyxl==3.1.5

